Refactor runner, output, and logging.

+++ lib/reviewer/history.rb
@@ -26,6 +26,8 @@ module Reviewer
     end
 
     def reset!
+      return unless File.exist?(file)
+
       FileUtils.rm(file)
     end
   end
